requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=py39haa95532_0
anyio=4.2.0=py39haa95532_0
asgiref=3.8.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.2.0=pypi_0
beautifulsoup4=4.12.3=py39haa95532_0
blas=1.0=mkl
blinker=1.6.2=py39haa95532_0
bottleneck=1.3.7=py39h9128911_0
brotli-python=1.0.9=py39hd77b12b_8
build=1.2.2=pypi_0
ca-certificates=2024.7.2=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py39haa95532_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
chroma-hnswlib=0.7.6=pypi_0
chromadb=0.5.5=pypi_0
click=8.1.7=py39haa95532_0
colorama=0.4.6=py39haa95532_0
coloredlogs=15.0.1=pypi_0
cryptography=43.0.1=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecated=1.2.14=pypi_0
distro=1.9.0=py39haa95532_0
exceptiongroup=1.2.0=py39haa95532_0
faiss-cpu=1.8.0.post1=pypi_0
fastapi=0.115.0=pypi_0
filelock=3.16.1=pypi_0
flask=3.0.3=py39haa95532_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
google-auth=2.34.0=pypi_0
googleapis-common-protos=1.65.0=pypi_0
greenlet=3.1.0=pypi_0
grpcio=1.66.1=pypi_0
h11=0.14.0=py39haa95532_0
httpcore=1.0.2=py39haa95532_0
httptools=0.6.1=pypi_0
httpx=0.27.0=py39haa95532_0
huggingface-hub=0.25.0=pypi_0
humanfriendly=10.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py39haa95532_0
importlib-metadata=7.0.1=py39haa95532_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
itsdangerous=2.2.0=py39haa95532_0
jinja2=3.1.4=py39haa95532_0
jiter=0.5.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
kubernetes=30.1.0=pypi_0
langchain=0.3.0=pypi_0
langchain-community=0.3.0=pypi_0
langchain-core=0.3.1=pypi_0
langchain-openai=0.2.0=pypi_0
langchain-text-splitters=0.3.0=pypi_0
langsmith=0.1.122=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h2bbff1b_0
marshmallow=3.22.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.10=py39h827c3e9_0
mkl_random=1.2.7=py39hc64d2fc_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
numexpr=2.8.7=py39h2cd9be0_0
numpy=1.26.4=py39h055cbcc_0
numpy-base=1.26.4=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.19.2=pypi_0
openai=1.46.0=pypi_0
openssl=1.1.1w=h2bbff1b_0
opentelemetry-api=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.27.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.27.0=pypi_0
opentelemetry-instrumentation=0.48b0=pypi_0
opentelemetry-instrumentation-asgi=0.48b0=pypi_0
opentelemetry-instrumentation-fastapi=0.48b0=pypi_0
opentelemetry-proto=1.27.0=pypi_0
opentelemetry-sdk=1.27.0=pypi_0
opentelemetry-semantic-conventions=0.48b0=pypi_0
opentelemetry-util-http=0.48b0=pypi_0
orjson=3.10.7=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=py39h5da7b33_0
pdfminer-six=20240706=pypi_0
pip=24.2=py39haa95532_0
posthog=3.6.6=pypi_0
protobuf=4.25.4=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pybind11-abi=5=hd3eb1b0_0
pycparser=2.22=pypi_0
pydantic=2.8.2=py39haa95532_0
pydantic-core=2.20.1=py39hefb1915_0
pydantic-settings=2.5.2=pypi_0
pypdf=5.0.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pyreadline3=3.5.3=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.6=h6244533_1
python-dateutil=2.9.0post0=py39haa95532_2
python-dotenv=1.0.1=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py39haa95532_0
pywin32=306=pypi_0
pyyaml=6.0.2=pypi_0
regex=2024.9.11=pypi_0
requests=2.32.3=py39haa95532_0
requests-oauthlib=2.0.0=pypi_0
rich=13.8.1=pypi_0
rsa=4.9=pypi_0
scipy=1.13.1=py39h8640f81_0
selenium=3.141.0=py39h2bbff1b_1000
setuptools=72.1.0=py39haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py39haa95532_0
soupsieve=2.5=py39haa95532_0
sqlalchemy=2.0.35=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.38.5=pypi_0
sympy=1.13.2=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tokenizers=0.20.0=pypi_0
tomli=2.0.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=py39h9909e9c_0
typer=0.12.5=pypi_0
typing-extensions=4.11.0=py39haa95532_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=py39haa95532_0
tzdata=2024a=h04d1e81_0
urllib3=2.0.3=py39haa95532_0
uvicorn=0.30.6=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
watchfiles=0.24.0=pypi_0
websocket-client=1.8.0=pypi_0
websockets=13.0.1=pypi_0
werkzeug=3.0.3=py39haa95532_0
wheel=0.44.0=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.16.0=pypi_0
yarl=1.11.1=pypi_0
zipp=3.17.0=py39haa95532_0
